Combine editor navigation test scenes
